requirements.txt
alembic==1.2.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
environs==6.0.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
ndg-httpsclient==0.5.1
pusher==2.1.4
pyasn1==0.4.7
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
